Add files via upload
The black-browed albatross (Thalassarche melanophris), also known as the black-browed mollymawk,[3] is a large seabird of the albatross family Diomedeidae; it is the most widespread and common member of its family.
Mollymawks are albatrosses in the family Diomedeidae and order Procellariiformes, which also includes shearwaters, fulmars, storm petrels, and diving petrels. These birds share certain identifying features. They have nasal passages that attach to the upper bill called naricorns, although the nostrils on the albatross are on the sides of the bill. The bills of Procellariiformes are also unique in that they are split into between seven and nine horny plates. They produce a stomach oil made up of wax esters and triglycerides that is stored in the proventriculus. This is used against predators as well as being an energy-rich food source for chicks and also for the adults during their long flights.[4] The albatross also has a salt gland above the nasal passage which helps to remove salt from the ocean water that they imbibe. The gland excretes a high saline solution through the bird's nose.