Add files via upload
The black-browed albatross (Thalassarche melanophris), also known as the black-browed mollymawk, is a large seabird of the albatross family Diomedeidae; it is the most widespread and common member of its family. Mollymawks are albatrosses in the family Diomedeidae and order Procellariiformes, which also includes shearwaters, fulmars, storm petrels, and diving petrels. These birds share certain identifying features. They have nasal passages that attach to the upper bill called naricorns, although the nostrils on the albatross are on the sides of the bill.