Add files via upload
The crab-plover or crab plover (Dromas ardeola) is a bird related to the waders, but sufficiently distinctive to merit its own family Dromadidae. Its relationship within the Charadriiformes is unclear, some have considered it to be closely related to the thick-knees, or the pratincoles, while others have considered it closer to the auks and gulls. It is the only member of the genus Dromas and is unique among waders in making use of ground warmth to aid incubation of the eggs.