Add files via upload
The Madagascan pygmy kingfisher (Corythornis madagascariensis) is a species of bird in the family Alcedinidae. It is endemic to Madagascar and found in western dry deciduous forests.

The first formal description of the Madagascan pygmy kingfisher was by the Swedish naturalist Carl Linnaeus in 1758 in the twelfth edition of his Systema Naturae under the binomial name Alcedo madagascariensis.

Although it has been placed in the genera Ceyx and Ispidina in the past, a study of its molecular phylogenetic relationships with other genera in the region suggests it is best treated as a member of the genus Corythornis.